extract features and remove dependency between skia and skia_c_bindings
