+++ skia-bindings/build_support.rs
@@ -5,6 +5,7 @@ pub mod cargo;
 pub mod clang;
 pub mod ios;
 pub mod llvm;
+pub mod features;
 pub mod skia;
 pub mod skia_c_bindings;
 pub mod vs;
