+++ skia-bindings/build_support/skia.rs
@@ -1,6 +1,6 @@
 //! Full build support for the Skia library, SkiaBindings library and bindings.rs file.
 
-use crate::build_support::{android, cargo, clang, ios, llvm, vs};
+use crate::build_support::{android, cargo, clang, features, ios, llvm, vs};
 use std::env;
 use std::path::{Path, PathBuf};
 use std::process::{Command, Stdio};
@@ -12,23 +12,9 @@ mod lib {
     pub const SK_PARAGRAPH: &str = "skparagraph";
 }
 
-/// Feature identifiers define the additional configuration parts of the binaries to download.
-mod feature_id {
-    pub const GL: &str = "gl";
-    pub const VULKAN: &str = "vulkan";
-    pub const METAL: &str = "metal";
-    pub const D3D: &str = "d3d";
-    pub const TEXTLAYOUT: &str = "textlayout";
-    pub const WEBPE: &str = "webpe";
-    pub const WEBPD: &str = "webpd";
-    pub const EGL: &str = "egl";
-    pub const X11: &str = "x11";
-    pub const WAYLAND: &str = "wayland";
-}
-
 /// The defaults for the Skia build configuration.
-impl Default for BuildConfiguration {
-    fn default() -> Self {
+impl From<features::Features> for BuildConfiguration {
+    fn from(features: features::Features) -> Self {
         let skia_debug = matches!(cargo::env_var("SKIA_DEBUG"), Some(v) if v != "0");
 
         BuildConfiguration {
@@ -36,21 +22,7 @@ impl Default for BuildConfiguration {
             skia_debug,
             // `OPT_LEVEL` is set by Cargo itself.
             opt_level: cargo::env_var("OPT_LEVEL"),
-            features: Features {
-                gl: cfg!(feature = "gl"),
-                egl: cfg!(feature = "egl"),
-                wayland: cfg!(feature = "wayland"),
-                x11: cfg!(feature = "x11"),
-                vulkan: cfg!(feature = "vulkan"),
-                metal: cfg!(feature = "metal"),
-                d3d: cfg!(feature = "d3d"),
-                text_layout: cfg!(feature = "textlayout"),
-                webp_encode: cfg!(feature = "webp-encode"),
-                webp_decode: cfg!(feature = "webp-decode"),
-                animation: false,
-                dng: false,
-                particles: false,
-            },
+            features,
             cc: cargo::env_var("CC").unwrap_or_else(|| "clang".to_string()),
             cxx: cargo::env_var("CXX").unwrap_or_else(|| "clang++".to_string()),
         }
@@ -71,99 +43,13 @@ pub struct BuildConfiguration {
     skia_debug: bool,
 
     /// The Skia feature set to compile.
-    pub features: Features,
+    features: features::Features,
 
     /// C compiler to use
     cc: String,
 
     /// C++ compiler to use
     cxx: String,
-}
-
-#[derive(Clone, PartialEq, Eq, Debug)]
-pub struct Features {
-    /// Build with OpenGL support?
-    pub gl: bool,
-
-    /// Build with EGL support? If you set X11, setting this to false will use LibGL (GLX)
-    pub egl: bool,
-
-    /// Build with Wayland support? This requires EGL, as GLX does not work on Wayland.
-    pub wayland: bool,
-
-    /// Build with X11 support?
-    pub x11: bool,
-
-    /// Build with Vulkan support?
-    pub vulkan: bool,
-
-    /// Build with Metal support?
-    pub metal: bool,
-
-    /// Build with Direct3D support?
-    pub d3d: bool,
-
-    /// Features related to text layout. Modules skshaper and skparagraph.
-    pub text_layout: bool,
-
-    /// Support the encoding of bitmap data to the WEBP image format.
-    pub webp_encode: bool,
-
-    /// Support the decoding of the WEBP image format to bitmap data.
-    pub webp_decode: bool,
-
-    /// Build with animation support (yet unsupported, no wrappers).
-    pub animation: bool,
-
-    /// Support DNG file format (currently unsupported because of build errors).
-    pub dng: bool,
-
-    /// Build the particles module (unsupported, no wrappers).
-    pub particles: bool,
-}
-
-impl Features {
-    pub fn gpu(&self) -> bool {
-        self.gl || self.vulkan || self.metal || self.d3d
-    }
-
-    /// Feature Ids used to look up prebuilt binaries.
-    pub fn ids(&self) -> Vec<&str> {
-        let mut feature_ids = Vec::new();
-
-        if self.gl {
-            feature_ids.push(feature_id::GL);
-        }
-        if self.egl {
-            feature_ids.push(feature_id::EGL);
-        }
-        if self.x11 {
-            feature_ids.push(feature_id::X11);
-        }
-        if self.wayland {
-            feature_ids.push(feature_id::WAYLAND);
-        }
-        if self.vulkan {
-            feature_ids.push(feature_id::VULKAN);
-        }
-        if self.metal {
-            feature_ids.push(feature_id::METAL);
-        }
-        if self.d3d {
-            feature_ids.push(feature_id::D3D);
-        }
-        if self.text_layout {
-            feature_ids.push(feature_id::TEXTLAYOUT);
-        }
-        if self.webp_encode {
-            feature_ids.push(feature_id::WEBPE);
-        }
-        if self.webp_decode {
-            feature_ids.push(feature_id::WEBPD);
-        }
-
-        feature_ids
-    }
 }
 
 /// This is the final, low level build configuration.
